refactor: navigtion 제네릭 타입 설정
