+++ src/shared/types/navigation.ts
@@ -1,0 +1,13 @@
+import { NativeStackNavigationProp } from '@react-navigation/native-stack';
+
+type RootStackParamList = {
+  Login: undefined;
+  Main: undefined;
+  Setting: undefined;
+  RecipeCreateForm: undefined;
+};
+
+export type RootNavigationProp<T extends keyof RootStackParamList> = NativeStackNavigationProp<
+  RootStackParamList,
+  T
+>;
